+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,9 +2,8 @@
 
 ## Unreleased
 
-* Update `OpenTelemetry.Api.ProviderBuilderExtensions` to `1.8.1`.
-  * Update `OpenTelemetry.Api` to `1.8.1`.
-  ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))
+* Updated OpenTelemetry core component version(s) to `1.9.0`.
+  ([#1888](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1888))
 
 ## 1.8.1
 
@@ -446,9 +445,11 @@ Released 2021-Apr-23
 
 * Sanitize `http.url` attribute.
   ([#1961](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1961))
+
 * Added `RecordException` to HttpClientInstrumentationOptions and
   HttpWebRequestInstrumentationOptions which allows Exception to be reported as
   ActivityEvent.
+
 * Update `AddHttpClientInstrumentation` extension method for .NET Framework to
   use only use `HttpWebRequestInstrumentationOptions`
   ([#1982](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1982))
@@ -461,6 +462,7 @@ Released 2021-Mar-19
   and ActivityProcessors. Samplers, ActivityProcessor.OnStart will now get the
   Activity before any enrichment done by the instrumentation.
   ([#1836](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1836))
+
 * Performance optimization by leveraging sampling decision and short circuiting
   activity enrichment.
   ([#1894](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1894))
@@ -495,12 +497,15 @@ Released 2020-Nov-5
   `HttpWebRequest` in Activity.CustomProperty. To enrich activity, use the
   Enrich action on the instrumentation.
   ([#1407](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1407))
+
 * Renamed TextMapPropagator to TraceContextPropagator, CompositePropagator to
   CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator and
   changed from interface to abstract class.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
+
 * Propagators.DefaultTextMapPropagator will be used as the default Propagator
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1428))
+
 * Removed Propagator from Instrumentation Options. Instrumentation now always
   respect the Propagator.DefaultTextMapPropagator.
   ([#1448](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1448))
@@ -513,6 +518,7 @@ Released 2020-Oct-16
   Activity.CustomProperty. To enrich activity, use the Enrich action on the
   instrumentation.
   ([#1261](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1261))
+
 * Span Status is populated as per new spec
   ([#1313](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1313))
 
@@ -529,6 +535,7 @@ Released 2020-08-28
 * HttpClient/HttpWebRequest instrumentation will now add the raw Request,
   Response, and/or Exception objects to the Activity it creates
   ([#1099](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1099))
+
 * Changed the default propagation to support W3C Baggage
   ([#1048](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1048))
   * The default ITextFormat is now `CompositePropagator(TraceContextFormat,
@@ -536,10 +543,12 @@ Released 2020-08-28
     Baggage](https://github.com/w3c/baggage/blob/master/baggage/HTTP_HEADER_FORMAT.md)
     header. Previously Baggage was sent using the `Correlation-Context` header,
     which is now outdated.
+
 * Removed `AddHttpInstrumentation` and `AddHttpWebRequestInstrumentation` (.NET
   Framework) `TracerProviderBuilderExtensions`. `AddHttpClientInstrumentation`
   will now register `HttpClient` instrumentation on .NET Core and `HttpClient` +
   `HttpWebRequest` instrumentation on .NET Framework.
+
 * Renamed `ITextPropagator` to `IPropagator`
   ([#1190](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1190))